Se realiza un refactoring para separar unas queries que estaban dentro de la carpeta commands y ser coherentes con CQRS
