+++ Backend/catalog/modules/catalog/application/queries/__init__.py
@@ -1,0 +1,1 @@
+# Paquete de queries del catálogo - patrón CQRS
