+++ Backend/catalog/modules/catalog/application/queries/obtain_property_by_category_id.py
@@ -2,7 +2,7 @@ from uuid import UUID
 
 
 class ObtainPropertyByCategoryId:
-	"""Comando para obtener la propiedad asociada a un id de categoria."""
+	"""Query para obtener la propiedad asociada a un UUID de categoria."""
 
 	def __init__(self, repository):
 		self.repository = repository
